+++ src/main/java/com/example/jpokebattle/gui/data/FaintedViewData.java
@@ -1,11 +1,9 @@
 package com.example.jpokebattle.gui.data;
 
-import com.example.jpokebattle.poke.Pokemon;
-
 public class FaintedViewData implements DynamicViewData {
     private final String faintedPokemon;
     private final String playerPokemon;
-    private final double expGained;
+    private final int expGained;
     private final boolean isPlayer;
 
     // Used when the player's pokemon faints
@@ -17,14 +15,14 @@ public class FaintedViewData implements DynamicViewData {
     }
 
     // Used when the enemy's pokemon faints
-    public FaintedViewData(String faintedPokemon, String playerPokemon, double expGained) {
+    public FaintedViewData(String faintedPokemon, String playerPokemon, int expGained) {
         this.faintedPokemon = faintedPokemon;
         this.playerPokemon = playerPokemon;
         this.expGained = expGained;
         this.isPlayer = false;
     }
 
-    public double getExpGained() { return expGained; }
+    public int getExpGained() { return expGained; }
     public String getFaintedPokemon() { return faintedPokemon; }
     public String getPlayerPokemon() { return playerPokemon; }
     public boolean isPlayer() { return isPlayer; }
